backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
pydantic-settings==2.5.2
python-multipart==0.0.9
openpyxl==3.1.5
pandas==2.2.2
